Revision (#1)
Substantial revision to the code for both bug fixes and quality of life improvements. 

Bugs
_____

Ellipse hard coding: When an ellipse was specified, the code was ignoring the two inputs for defining the height and width of the ellipse and instead randomly setting its size.

Run number: The number of runs being simulated was less than the number of runs specified, due to an indexing error.

Improvements
_____

Signal magnitude: signal magnitude is now being computed using sim.compute_signal_change which offers more control than before

Smoothing: Added functionality for smoothing the brain before averaging/deconvolution

Notebook: Vastly increased the amount of detail in the notebook and created some warnings when the files do not exist. Also offers some example code and adds more plots that can be generated.

Comments: More comments are added throughout to make it easier to follow.