Revision_2
Substantial revision to the code to enable using different distance metrics, testing the effects, and quality of life improvements.

Improvements
_____

Enable different distance metrics (e.g. correlation) to be used for simulating data

Added reliability tests to the pipeline using FSL's randomise function

Notebook is updated to focus on the main results

Increased documentation in the notebook and improved Table of Contents